submissions/team-members/yan-cotta/requirements.txt
# Contains all libraries that are needed for this project

Pillow
kaggle
matplotlib
numpy
split-folders
torch
torchvision